requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.1
requests==2.31.0
openai==1.12.0
